requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
cffi==1.15.1
cryptography==38.0.4
Django==4.1.3
django-cors-headers==3.13.0
django-rest-knox==4.2.0
djangorestframework==3.14.0
gunicorn==20.1.0
pycparser==2.21
python-decouple==3.6
pytz==2022.6
sqlparse==0.4.3
whitenoise==6.2.0
